Move mock out of tests folder
